cpu_requirements.txt
torch==2.2.2
torchvision==0.17.2
flask==2.3.3
flasgger==0.9.7.1
Pillow==10.2.0
matplotlib==3.8.3
scikit-learn==1.4.1.post1
